pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager"
authors = ["Dmitri Spirin <yamsky.xbox@gmail.com>"]
packages = [
  { include = "task_manager" },
]


[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.6"
whitenoise = "^6.2.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
coverage = "^6.4.2"
django-bootstrap4 = "^22.1"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
